backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
Werkzeug==3.0.1

# Security and validation
python-dotenv==1.0.0
requests==2.31.0

# Audio processing
pydub==0.25.1
numpy==1.26.2
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1

# AI and external APIs
openai==1.3.8

# Production utilities
redis==5.0.1
celery==5.3.4
psutil==5.9.6

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Authentication and security
PyJWT==2.8.0
bcrypt==4.1.2
Flask-Limiter==3.5.0